requirements.txt
astrapy
mcp
fastmcp
cassandra-driver
cassio
fastmcp
gitpython
radon
tree-sitter
unstructured
twilio
fastapi
uvicorn
httpx
python-dotenv
